Update profile to follow latest FBPIC format
