+++ fbpic/lpa_utils/laser/laser_profiles.py
@@ -835,223 +835,3 @@ class FewCycleLaser( LaserProfile ):
         Ey = self.E0y * profile
 
         return( Ex.real, Ey.real )
-
-
-class CustomSpectrumLaserPulse( LaserProfile ):
-    """Class that calculates a Gaussian laser pulse with a user defined
-    spectral amplitude and phase."""
-
-    def __init__( self, a0, waist, z0, spectrum_file, zf=None,
-                    theta_pol=0., propagation_direction=1 ):
-        """
-        Define a linearly-polarized Gaussian laser profile with a
-        user-provided temporal profile (defined via spectral phase and amplitude).
-
-        The spatial profile of the electric field **near the focal plane**
-        is given by:
-
-        .. math::
-
-            E(\\boldsymbol{x},t) = a_0\\times E_0\,
-            \exp\left( -\\frac{r^2}{w_0^2} \\right)
-
-        where :math:`E_0 = m_e c^2 k_0 / q_e` is the field amplitude for :math:`a_0=1`.
-
-        .. note::
-
-            The additional terms that arise **far from the focal plane**
-            (Gouy phase, wavefront curvature, ...) are not included in the above
-            formula for simplicity, but are of course taken into account by
-            the code, when initializing the laser pulse away from the focal plane.
-
-        The temporal characteristics of the pulse are calculated numerically via the
-        spectral phase and amplitude which are provided to the class as a path to
-        a csv file containing the data.
-
-        Parameters
-        ----------
-
-        a0: float (dimensionless)
-            The peak normalized vector potential at the focal plane, defined
-            as :math:`a_0` in the above formula.
-
-        waist: float (in meter)
-            Laser waist at the focal plane, defined as :math:`w_0` in the
-            above formula.
-
-        z0: float (in meter)
-            The initial position of the centroid of the laser
-            (in the lab frame), defined as :math:`z_0` in the above formula.
-
-        spectrum_file: file path
-            The path to a csv file containing 3 columns (no headers). The three columns
-            should represent wavelength (in m), spectral amplitude (arb. units)
-            and spectral phase (in radians). Use a "\t" tab as the deliminator in the file.
-
-        zf: float (in meter), optional
-            The position of the focal plane (in the lab frame).
-            If ``zf`` is not provided, the code assumes that ``zf=z0``, i.e.
-            that the laser pulse is at the focal plane initially.
-
-        theta_pol: float (in radian), optional
-           The angle of polarization with respect to the x axis.
-
-
-        propagation_direction: int, optional
-            Indicates in which direction the laser propagates.
-            This should be either 1 (laser propagates towards positive z)
-            or -1 (laser propagates towards negative z).
-        """
-
-       # Initialize propagation direction and mark the profile as GPU capable
-        LaserProfile.__init__(self, propagation_direction, gpu_capable=True)
-
-        # Import the laser temporal profile as defined by the user
-        self.spectrum_file    = spectrum_file
-        t_user, Et_user, lambda0 = self.create1DPulseFromSpec()
-        self.t_user = t_user
-        self.Et_user = Et_user
-
-        # Set a number of parameters for the laser
-        k0 = 2*np.pi/lambda0
-        E0 = a0*m_e*c**2*k0/e
-        zr = 0.5*k0*waist**2
-
-        # If no focal plane position is given, use z0
-        if zf is None:
-            zf = z0
-
-        # Store the parameters
-        self.k0 = k0
-        self.inv_zr = 1./zr
-        self.zf = zf
-        self.z0 = z0
-        self.E0x = E0 * np.cos(theta_pol)
-        self.E0y = E0 * np.sin(theta_pol)
-        self.w0 = waist
-
-
-    def E_field( self, x, y, z, t ):
-        """
-        See the docstring of LaserProfile.E_field
-        """
-
-        # Diffraction factor
-        prop_dir = self.propag_direction
-        diffract_factor = 1. + 1j * prop_dir*(z - self.zf) * self.inv_zr
-
-        # Spatial grid
-        z_ax = z[:, 0, 0]
-
-        # Interpolate the temporal profile of the pulse onto this axis.
-        # we center the pulse temporally around the pulse starting point
-        fpulse = interp1d(c*self.t_user-self.z0,self.Et_user,fill_value=0,bounds_error=False)
-        timeProfile = fpulse(z_ax)
-
-        # And apply it to the full z-grid
-        long_profile = np.zeros(z.shape, dtype=np.complex64)
-        for i in range(len(timeProfile)):
-            long_profile[i, :, :] = timeProfile[i]
-
-        # Calculate the normalised transverse profile
-        trans_profile = np.exp( - (x**2 + y**2) / (self.w0**2 * diffract_factor))/diffract_factor
-
-        # Get the projection along x and y, with the correct polarization
-        Ex = self.E0x * long_profile * trans_profile
-        Ey = self.E0y * long_profile * trans_profile
-
-        return( Ex.real, Ey.real )
-
-
-    def create1DPulseFromSpec(self):
-
-        """ Taking a spectrum file containing tab seperated columns with wavelength (m) spectral Intensity (arb)
-        and spectral phase (rad) we calculate the normalized temporal profile of the laser pulse
-        """
-
-        # Read in the Spectrum File and put the data in arrays
-        f = open(self.spectrum_file)
-        wavelength = []
-        intensity = []
-        phase = []
-        for line in f.readlines():
-            vals = line.split('\t')
-            wavelength.append(float(vals[0]))
-            intensity.append(float(vals[1]))
-            phase.append(float(vals[2]))
-
-        wavelength = np.asarray(wavelength)
-        intensity = np.asarray(intensity)
-        phase = np.asarray(phase)
-
-        # central wavelength
-        lambda0 = np.trapz(wavelength*intensity,wavelength)/np.trapz(intensity,wavelength)
-        omega0 = 2*np.pi*c/lambda0
-
-        # Computation Parameters
-        lambda_resolution = lambda0/1000           # spectral resolution defined by wavelength
-        dt                = lambda0/c/1000         # temporal resolution defined as fraction of an optical cycle
-        time_window       = lambda0 * lambda0 / c / lambda_resolution
-        wavelength_range  = lambda0 * lambda0 / c / dt
-        Nt                = np.round(time_window/dt).astype(int)
-        Nlambda           = np.round(wavelength_range/lambda_resolution).astype(int)
-
-        # Define the time array and its corresponding frequency array after a FT
-        time_arr          = np.linspace(-time_window/2,(time_window/2-dt),Nt)
-        freq_arr, _        = self.FFT(time_arr, np.zeros_like(time_arr))
-
-        # Interpolate the user defined spectral amplitude and phase onto the new frequency
-        # array.
-        spectral_inten_fn = interp1d(2*np.pi*c/wavelength,intensity,fill_value=0,bounds_error=False)
-        spectral_phase_fn = interp1d(2*np.pi*c/wavelength,phase,fill_value=0,bounds_error=False)
-
-        # Calculate the normalised temporal profile of the electric field from user defined spectrum
-        _, Et = self.IFFT(freq_arr, np.sqrt(spectral_inten_fn(freq_arr))*np.exp(1j*spectral_phase_fn(freq_arr)))
-        Et = Et/np.max(np.real(Et))
-
-        return time_arr, Et, lambda0
-
-    def FFT(self,t, f):
-        """ Function to calculate the discrete approximation to the continuous fourier transform
-        according to:
-        $$
-        F(\omega) = \int^{\infty}_{-\infty} f(t) e^{xs-i \omega t} dt
-        $$
-
-        In this convetion, the inverse fourier transform is given by:
-        $$
-        f(t) = \frac{1}{2\pi} \int^{\infty}_{-\infty} F(\omega) e^{i \omega t} d\omega
-        $$
-
-
-        t is the independant variable
-        f is the function evaluated at t. eg. f(t)
-        """
-        dt = t[1]-t[0]
-        F = np.fft.fftshift(np.fft.fft(np.fft.ifftshift(f)))*dt
-        T = t[-1] - t[0]
-        sampleFreq = len(t)/T
-        omega = 2*np.pi*np.linspace(-sampleFreq/2, (sampleFreq/2-sampleFreq/len(t)) , len(t))
-        return (omega,F)
-
-    def IFFT(self,omega, F):
-        """ Function to calculate the discrete approximation to the continuous inverse fourier transform
-        according to:
-        $$
-        F(\omega) = \int^{\infty}_{-\infty} f(t) e^{-i \omega t} dt
-        $$
-
-        In this convetion, the inverse fourier transform is given by:
-        $$
-        f(t) = \frac{1}{2\pi} \int^{\infty}_{-\infty} F(\omega) e^{i \omega t} d\omega
-        $$
-
-
-        omega is the independant variable
-        F is the function evaluated at omega. eg. F(omega)
-        """
-        sampleFreq = len(omega)*abs(omega[3]-omega[2])/(2*np.pi)
-        dt = 1/sampleFreq
-        t = np.linspace(-dt*len(omega)/2 , dt*len(omega)/2 - dt, len(omega))
-        f = np.fft.fftshift(np.fft.ifft(np.fft.ifftshift(F)))/dt
-        return (t,f)
